Chore: 변수명 수정, STEP2 분리
부적절한 변수명들을 수정하고 추후 STEP3과제를 위해 STEP2와 main파일을 따로 구성함.
