+++ CodeStarterCamp_Week2/STEP3.swift
@@ -1,0 +1,8 @@
+//
+//  STEP3.swift
+//  CodeStarterCamp_Week2
+//
+//  Created by Minseok on 2022/04/23.
+//
+
+import Foundation
